Moving Login from min_iteration2 to response_screen (and renaming min_iteration2 to response_screen)
