+++ groupformer/urls.py
@@ -18,7 +18,7 @@ from django.urls import include, path
 urlpatterns = [
     path('setup_screen/', include('setup_screen.urls')),
     path('min_iteration2/', include('min_iteration2.urls')),
-    path('min_iteration3/', include('min_iteration3.urls')),
+    path('response_screen/', include('response_screen.urls')),
     path('admin/', admin.site.urls),
     path('dbtools/', include('dbtools.urls')),
 ]
