+++ response_screen/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class MinIteration3Config(AppConfig):
-    name = 'min_iteration3'
+    name = 'response_screen'
