+++ response_screen/static/main/style.css
