+++ response_screen/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import get_object_or_404, render
+from django.shortcuts import get_object_or_404, render, redirect
 from django.http import JsonResponse, HttpResponse
 from django.template import RequestContext
 from django.urls import reverse
@@ -16,7 +16,7 @@ def response_screen(request, groupformer_id):
         participants = Participant.objects.filter(group_former=groupformer_id)
         context = {"projects": projects, "attributes": attributes, "participants": participants}
 
-        return render(request, 'min_iteration3_main/response_screen.html', context)
+        return render(request, 'response_screen_main/response_screen.html', context)
 
     return HttpResponse("404", status=404)
 
@@ -26,7 +26,7 @@ def groupformer_list(request):
     groupformers = GroupFormer.objects.all()
     context = {"groupformers": groupformers}
 
-    return render(request, 'min_iteration3_main/groupformer_list.html', context)
+    return render(request, 'response_screen_main/groupformer_list.html', context)
 
 
 def sample_groups(request, groupformer_id):
@@ -55,3 +55,20 @@ def sample_groups(request, groupformer_id):
         return JsonResponse({"data":sections[(groupformer_id - 1) % 2]})
     else:
         return JsonResponse({"data":[]}, status=404)
+
+
+def login_group(request, groupformer_id):
+    gfs = GroupFormer.objects.filter(pk=groupformer_id)
+    if len(gfs) == 0:
+        return render(request, 'response_screen_main/loginerror.html')
+    gf = gfs[0]
+
+    if request.POST.get("email"):
+        # Validate the login
+        parts = gf.getParticipantByEmail(request.POST["email"])
+        if parts == None:
+            return render(request, 'response_screen_main/login.html', {"groupformer": gf, 'error': True})
+        return redirect(reverse('reverse:response_screen', kwargs={"groupformer_id": gf.pk}))
+
+    # Log into the groupformer for the first time
+    return render(request, 'response_screen_main/login.html', {"groupformer": gf})
